requirements.txt
python-dotenv
aio-pika
boto3
pillow